型の変更: CGrepAgent.cpp、CSearchAgent.cpp、CSearchAgent.h、CLaxInteger.h、CStrictInteger.h、primitive.h、SakuraBasis.h、各変換クラスにおいて、位置を示す引数やメンバ変数の型をint型からssize_t型に変更し、大容量データの処理精度を向上させる修正を実施。

+++ sakura_core/agent/CGrepAgent.cpp
@@ -375,7 +375,7 @@ DWORD CGrepAgent::DoGrep(
 	CBregexp	cRegexp;
 	CNativeW	cmemMessage;
 	CNativeW	cUnicodeBuffer;
-	int			nWork;
+	ssize_t		nWork;
 	SGrepOption	sGrepOption;
 
 	/*
@@ -1521,7 +1521,7 @@ int CGrepAgent::DoGrepFile(
 			if( sSearchOption.bRegularExp ){
 				ssize_t nIndex = 0;
 #ifdef _DEBUG
-				int nIndexPrev = -1;
+				ssize_t nIndexPrev = -1;
 #endif
 
 				//	Jun. 21, 2003 genta ループ条件見直し
