+++ sakura_core/convert/CConvert_TabToSpace.cpp
@@ -22,21 +22,21 @@ bool CConvert_TabToSpace::DoConvert(CNativeW* pcData)
 	std::wstring buffer;
 	buffer.reserve(source.length() + numOfTabs * (m_nTabWidth - 1));
 
-	int begin = 0;
+	ssize_t begin = 0;
 	while (true) {
-		int lineLength;
+		ssize_t lineLength;
 		CEol eol;
 		/* CRLFで区切られる「行」を返す。CRLFは行長に加えない */
 		const wchar_t* line = GetNextLineW(source.data(),
-			static_cast<int>(source.length()), &lineLength, &begin, &eol, m_bExtEol);
+			source.length(), &lineLength, &begin, &eol, m_bExtEol);
 		if (!line)
 			break;
 		// 先頭行については開始桁位置を考慮する（さらに折り返し関連の対策が必要？）
-		int pos = (source.data() == line) ? m_nStartColumn : 0;
-		for (int i = 0; i < lineLength; ++i) {
+		ssize_t pos = (source.data() == line) ? m_nStartColumn : 0;
+		for (ssize_t i = 0; i < lineLength; ++i) {
 			if (line[i] == L'\t') {
-				const int width = m_nTabWidth - (pos % m_nTabWidth);
-				buffer.append(width, L' ');
+				const ssize_t width = m_nTabWidth - (pos % m_nTabWidth);
+				buffer.append(static_cast<size_t>(width), L' ');
 				pos += width;
 			} else {
 				buffer.push_back(line[i]);
