+++ sakura_core/convert/CConvert_ToZenhira.cpp
@@ -18,7 +18,7 @@ bool CConvert_ToZenhira::DoConvert(CNativeW* pcData)
 {
 	//半カナ→全角カナ
 	wchar_t* pBuf = new wchar_t[pcData->GetStringLength()+1]; //文字数が減ることはあっても増えることは無いので、これでＯＫ
-	int nBufLen = 0;
+	ssize_t nBufLen = 0;
 	Convert_HankataToZenkata(pcData->GetStringPtr(), pcData->GetStringLength(), pBuf, &nBufLen);
 
 	//全カナ→全角ひらがな
